+++ src/components/signin/SigninPage.test.js
@@ -1,21 +1,70 @@
 import React from 'react';
 import SignInPage from './SigninPage';
 
-describe('Sign in page', () => {
+describe('Sign in form test', () => {
+    let mountedSignInForm = null,
+       props = null;
     const mockHandleSubmitForm = jest.fn(),
         mockInputChange = jest.fn();
-    const component = shallow(
-        <SignInPage
-            handleLoginInputChange={mockInputChange}
-            handleLoginSubmitForm={mockHandleSubmitForm}
-        />
-    );
-    it('renders sign in page without crashing', () => {
-        expect(component).toMatchSnapshot();
+    const component = () => {
+        if (!mountedSignInForm) {
+            mountedSignInForm = mount(
+                <SignInPage {...props} />
+            );
+        }
+
+        return mountedSignInForm;
+    };
+
+    beforeEach(() => {
+        props = {
+            'handleLoginInputChange': mockHandleSubmitForm,
+            'handleLoginSubmitForm': mockInputChange,
+            'loginPutErrorMessage': null,
+            'loginSuccessMessage': null
+        };
+        mountedSignInForm = null;
     });
 
-    it('rendered `SignInForm` component receive four props', () => {
-        const signupFormDisplay = component.find('SignInForm');
-        expect(Object.keys(signupFormDisplay.props()).length).toBe(4);
+    it('renders and matches our snapshot', () => {
+        expect(component()).toMatchSnapshot();
     });
+
+    it('contains the form', () => {
+        expect(component().find('input')).toHaveLength(2);
+        expect(component().find('button')).toHaveLength(1);
+    });
+
+    it('should call `handleLoginInputChange` when input field is changed', () => {
+        component().find('input.test').simulate('change');
+        expect(props.handleLoginInputChange).toBeCalled();
+    });
+
+    it('should call `handleLoginSubmitForm` when submit button is called', () => {
+        component().find('form').simulate('submit');
+        expect(props.handleLoginSubmitForm).toBeCalled();
+    });
+
+    describe('when `loginSuccessMessage` is passed', () => {
+        beforeEach(() => {
+            props.loginSuccessMessage = 'You logged in successfully!';
+        });
+
+        it('renders a `Success Alert box`', () => {
+            const alertDiv = component().find('div.alert-success');
+            expect(alertDiv.length).toBe(1);
+        });
+    });
+
+    describe('when `loginPutErrorMessage` is passed', () => {
+        beforeEach(() => {
+            props.loginPutErrorMessage = 'Invalid email or password!';
+        });
+
+        it('renders a `Error Alert box`', () => {
+            const alertDiv = component().find('div.alert-danger');
+            expect(alertDiv.length).toBe(1);
+        });
+    });
+
 });
